telegram-bot/requirements.txt
python-telegram-bot==21.0
httpx~=0.27
python-dotenv==1.0.0
